Improve the expressivity of the ComputeQualityIndicators class.
Rather than missing the type of solution, we request it as a generic.
We can do it because the instance builds on a single experiment, so
the type of solution depends on it and never changes. The advantage
of doing so is that now we know the type (S) and so we can fake a cast
when using the indicator. Consequently, we fix all raw type
warnings by adding only a single cast warning, which is by the way
exactly where the problem is: at the misalignment between indicator
and population. So it makes a better indicator.