Weight vectors are now an attribute of the WASFGA class. Thus GWASFGA, which extends WASFGA, can use this new attribute to store its weight vectors.
In addition, we have fixed the usage of the weightVectorsFileName parameter. Now, if this parameter is not given, WASFGA generates the weight vectors (for two objectives problems) or read them from the resources folder of jMetal (for three or more objectives). Contrarily, if the weightVectorsFileName parameter is given, weight vectors are read from that file.