Remove useless life time of random generators.
Randomization is only used at instantiation time, thus it does not make
sense to retain this instance during all the life time of the solution.
Not only we save significant memory space by removing a field in all
solution instances, but we also centralize the responsibilities where
relevant, which unlocks future refactoring and optimizations.

The now obsolete field is only deprecated to not break existing
implementations. It should be removed for the next non-retrocompatible
release.