Factor copy constructors to AbstractGenericSolution.
Since it is highly similar to the problem-only constructor, both are
further factored into a more generic constructor based on variables and
objectives providers + attributes.

It also allows to provide an intermediary constructor with custom variables
only, so I added it too since it is useful for DefaultDoubleBinarySolution
and DefaultIntegerDoubleSolution.

With these additional constructors, initializeObjectiveValues can be
properly integrated and is not needed anymore. I made it deprecated to not
break external dependencies.